+++ test/e2e/steps/update_app_version.dart
@@ -15,22 +15,16 @@
 // along with this program. If not, see
 // <https://www.gnu.org/licenses/agpl-3.0.html>.
 
-import 'package:flutter/widgets.dart' show PageController;
-import 'package:get/get.dart';
+import 'package:gherkin/gherkin.dart';
+import 'package:messenger/config.dart';
 
-/// [StyleView] section.
-enum StyleTab { colors, typography, widgets }
+import '../world/custom_world.dart';
 
-/// Controller of a [StyleView].
-class StyleController extends GetxController {
-  /// Indicator whether the [Color]s of the [StyleView] should be inverted.
-  ///
-  /// Meant to be used as a light/dart theme switch.
-  final RxBool inverted = RxBool(false);
-
-  /// Selected [StyleTab].
-  final Rx<StyleTab> tab = Rx(StyleTab.colors);
-
-  /// [PageController] controlling the [PageView] of [StyleView].
-  final PageController pages = PageController();
-}
+/// Changes [Config.version] to a `test.0.0.1` string.
+///
+/// Examples:
+/// - When application version is updated
+final StepDefinitionGeneric updateAppVersion = then<CustomWorld>(
+  'application version is updated',
+  (context) async => Future.sync(() => Config.version = 'test.0.0.1'),
+);